requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
yt-dlp>=2025.1.15
notion-client==2.3.0
python-dotenv==1.0.1
supabase==2.10.0
youtube-transcript-api>=1.2.4
slowapi==0.1.9
limits==3.7.0
httpx>=0.27.0
PyJWT>=2.8.0
cryptography>=42.0.0
trafilatura>=2.0.0
PyMuPDF>=1.25.0
pdfminer.six>=20231228
